Extreme refactoring 🔥
