+++ flutter/lib/screens/student/components/student_course_item.dart
@@ -1,16 +1,12 @@
-import 'package:exams_quizzes_alike/models/available_exam.dart';
-import 'package:exams_quizzes_alike/models/exam.dart';
-import 'package:exams_quizzes_alike/screens/exam/exam_page.dart';
-import 'package:exams_quizzes_alike/screens/login/login_page.dart';
-import 'package:exams_quizzes_alike/screens/student/grade_exam/grade_exam_body.dart';
+import 'package:exams_quizzes_alike/models/course.dart';
+import 'package:exams_quizzes_alike/screens/student/course/student_course_page.dart';
 import 'package:exams_quizzes_alike/widgets/grid_item.dart';
 import 'package:flutter/material.dart';
 
-class StudentExamItem implements GridItem {
-  final Exam exam;
-  final AvailableExam availableExam;
+class StudentCourseItem implements GridItem {
+  final Course course;
 
-  StudentExamItem(this.exam, this.availableExam);
+  StudentCourseItem(this.course);
 
   @override
   Widget buildItem(BuildContext context) {
@@ -19,12 +15,16 @@ class StudentExamItem implements GridItem {
       child: InkWell(
         splashColor: Colors.white,
         onTap: () => Navigator.push(
-            context, MaterialPageRoute(builder: ((context) => LoginPage()))),
+            context,
+            MaterialPageRoute(
+                builder: ((context) => StudentCoursePage(
+                      course: course,
+                    )))),
         child: Center(
             child: Padding(
           padding: const EdgeInsets.all(5),
           child: Text(
-            exam.name + "\n\n" + "Due " + availableExam.maxDate,
+            course.courseDescription!,
             style: const TextStyle(
                 fontWeight: FontWeight.bold, fontSize: 15, color: Colors.white),
             textAlign: TextAlign.center,
