+++ flutter/lib/screens/student/components/student_exam_item.dart
@@ -1,13 +1,15 @@
-import 'package:exams_quizzes_alike/models/course.dart';
-import 'package:exams_quizzes_alike/screens/course/teacher_course_page.dart';
-import 'package:exams_quizzes_alike/screens/student/components/student_course_page.dart';
+import 'package:exams_quizzes_alike/models/available_exam.dart';
+import 'package:exams_quizzes_alike/models/exam.dart';
+import 'package:exams_quizzes_alike/screens/login/login_page.dart';
+import 'package:exams_quizzes_alike/screens/student/grade_exam/grade_exam_body.dart';
 import 'package:exams_quizzes_alike/widgets/grid_item.dart';
 import 'package:flutter/material.dart';
 
-class StudentCourseItem implements GridItem {
-  final Course course;
+class StudentExamItem implements GridItem {
+  final Exam exam;
+  final AvailableExam availableExam;
 
-  StudentCourseItem(this.course);
+  StudentExamItem(this.exam, this.availableExam);
 
   @override
   Widget buildItem(BuildContext context) {
@@ -16,16 +18,14 @@ class StudentCourseItem implements GridItem {
       child: InkWell(
         splashColor: Colors.white,
         onTap: () => Navigator.push(
+            //TODO PRESENT EXAM PAGE
             context,
-            MaterialPageRoute(
-                builder: ((context) => StudentCoursePage(
-                      course: course,
-                    )))),
+            MaterialPageRoute(builder: ((context) => LoginPage()))),
         child: Center(
             child: Padding(
           padding: const EdgeInsets.all(5),
           child: Text(
-            course.courseDescription!,
+            exam.name + "\n\n" + "Due " + availableExam.maxDate,
             style: const TextStyle(
                 fontWeight: FontWeight.bold, fontSize: 15, color: Colors.white),
             textAlign: TextAlign.center,
