+++ flutter/lib/screens/student/course/student_course_page.dart
@@ -1,5 +1,5 @@
 import 'package:exams_quizzes_alike/models/course.dart';
-import 'package:exams_quizzes_alike/screens/student/components/student_course_page_body.dart';
+import 'package:exams_quizzes_alike/screens/student/course/student_course_page_body.dart';
 import 'package:exams_quizzes_alike/widgets/main_app_bar.dart';
 import 'package:flutter/material.dart';
