+++ flutter/lib/screens/student/course/student_course_page_body.dart
@@ -3,7 +3,7 @@ import 'package:exams_quizzes_alike/models/course.dart';
 import 'package:exams_quizzes_alike/models/exam.dart';
 import 'package:exams_quizzes_alike/network/available_exam_requests.dart';
 import 'package:exams_quizzes_alike/network/exam_requests.dart';
-import 'package:exams_quizzes_alike/screens/student/components/exam/student_exam_item.dart';
+import 'package:exams_quizzes_alike/screens/student/components/student_exam_item.dart';
 import 'package:exams_quizzes_alike/screens/teacher/components/teacher_exam_item.dart';
 import 'package:flutter/material.dart';
