+++ flutter/lib/screens/teacher/components/teacher_course_item.dart
@@ -1,5 +1,5 @@
 import 'package:exams_quizzes_alike/models/course.dart';
-import 'package:exams_quizzes_alike/screens/course/teacher_course_page.dart';
+import 'package:exams_quizzes_alike/screens/teacher/course/teacher_course_page.dart';
 import 'package:exams_quizzes_alike/widgets/grid_item.dart';
 import 'package:flutter/material.dart';
