+++ flutter/lib/screens/teacher/exam/components/add_question/add_question_page.dart
@@ -1,5 +1,5 @@
 import 'package:exams_quizzes_alike/models/exam.dart';
-import 'package:exams_quizzes_alike/screens/exam/components/add_question/add_question_body.dart';
+import 'package:exams_quizzes_alike/screens/teacher/exam/components/add_question/add_question_body.dart';
 import 'package:exams_quizzes_alike/widgets/main_app_bar.dart';
 import 'package:flutter/material.dart';
