+++ flutter/lib/screens/teacher/exam/exam_page_body.dart
@@ -1,7 +1,7 @@
 import 'package:exams_quizzes_alike/models/exam.dart';
 import 'package:exams_quizzes_alike/models/question.dart';
 import 'package:exams_quizzes_alike/network/question_requests.dart';
-import 'package:exams_quizzes_alike/screens/exam/components/add_question/add_question_page.dart';
+import 'package:exams_quizzes_alike/screens/teacher/exam/components/add_question/add_question_page.dart';
 import 'package:exams_quizzes_alike/widgets/question_widget.dart';
 import 'package:flutter/material.dart';
