+++ flutter/lib/screens/teacher/teacher_page.dart
@@ -3,10 +3,10 @@ import 'package:exams_quizzes_alike/exceptions/exam_exception.dart';
 import 'package:exams_quizzes_alike/models/course.dart';
 import 'package:exams_quizzes_alike/network/course_requests.dart';
 import 'package:exams_quizzes_alike/network/exam_requests.dart';
-import 'package:exams_quizzes_alike/screens/exam/components/create/create_exam_page.dart';
 import 'package:exams_quizzes_alike/screens/teacher/components/teacher_course_item.dart';
 import 'package:exams_quizzes_alike/screens/teacher/components/teacher_exam_item.dart';
 import 'package:exams_quizzes_alike/screens/teacher/components/teacher_exam_item.dart';
+import 'package:exams_quizzes_alike/screens/teacher/exam/components/create/create_exam_page.dart';
 import 'package:exams_quizzes_alike/widgets/main_app_bar.dart';
 import 'package:flutter/material.dart';
